Start fixing errors from the requirement of the ui component needing to be declared in the infoFactory function.
