+++ inst/editor/src/ui-node-definitions/uiNodeTypes.ts
@@ -40,11 +40,9 @@ import { input_checkbox } from "./Shiny/input_checkbox";
 import { input_checkbox_group } from "./Shiny/input_checkbox_group";
 import { input_radio_buttons } from "./Shiny/input_radio_buttons";
 import { input_select } from "./Shiny/input_select";
-import { input_slider } from "./Shiny/input_slider";
 import { input_text } from "./Shiny/input_text";
 import { layout_sidebar } from "./Shiny/layout_sidebar";
 import { markdown_node } from "./Shiny/markdown";
-import { output_plot } from "./Shiny/output_plot";
 import { output_text } from "./Shiny/output_text";
 import { output_ui } from "./Shiny/output_ui";
 import { panel_main } from "./Shiny/panel_main";
@@ -61,9 +59,11 @@ import { shinyNavbarPageInfo } from "./Shiny/ShinyNavbarPage";
 import { shinyNumericInputInfo } from "./Shiny/ShinyNumericInput";
 import { input_numeric } from "./Shiny/ShinyNumericInput/input_numeric";
 import { shinyPlotOutputInfo } from "./Shiny/ShinyPlotOutput";
+import { output_plot } from "./Shiny/ShinyPlotOutput/output_plot";
 import { shinyRadioButtonsInfo } from "./Shiny/ShinyRadioButtons/ShinyRadioButtons";
 import { shinySelectInputInfo } from "./Shiny/ShinySelectInput";
 import { shinySliderInputInfo } from "./Shiny/ShinySliderInput";
+import { input_slider } from "./Shiny/ShinySliderInput/input_slider";
 import { shinyTabsetPanelInfo } from "./Shiny/ShinyTabsetPanel";
 import { shinyTextInputInfo } from "./Shiny/ShinyTextInput";
 import { shinyTextOutputInfo } from "./Shiny/ShinyTextOutput/ShinyTextOutput";
